# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ACCweb api documentation",
        "title": "ACCWeb API documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "ACCWeb project",
            "url": "https://github.com/assetto-corsa-web/accweb"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/assetto-corsa-web/accweb/blob/master/LICENSE"
        },
        "version": "1.19"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/instance": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Create new acc instance information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instances"
                ],
                "summary": "Create new acc instance information",
                "parameters": [
                    {
                        "description": "Instance data",
                        "name": "instance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.SaveInstancePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.InstancePayload"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.AccWError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.AccWError"
                        }
                    }
                }
            }
        },
        "/instance/{id}": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get acc instance information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instances"
                ],
                "summary": "Get acc instance information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Instance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.InstancePayload"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Saves acc instance information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instances"
                ],
                "summary": "Saves acc instance information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Instance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Instance data",
                        "name": "instance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.SaveInstancePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.InstancePayload"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.AccWError"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.AccWError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Delete acc instance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instances"
                ],
                "summary": "Delete acc instance",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Instance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.AccWError"
                        }
                    }
                }
            }
        },
        "/instance/{id}/clone": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Clones acc instance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instances"
                ],
                "summary": "Clones acc instance",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Instance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.AccWError"
                        }
                    }
                }
            }
        },
        "/instance/{id}/export": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get acc instance configuration files",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instances"
                ],
                "summary": "Get acc instance configuration files",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Instance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Zip file with all cofiguration files",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.AccWError"
                        }
                    }
                }
            }
        },
...
        "/instance/{id}/logs": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get acc instance logs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instances"
                ],
                "summary": "Get acc instance logs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Instance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.accWebInstanceLogs"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.AccWError"
                        }
                    }
                }
            }
        },
        "/instance/{id}/start": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Starts acc instance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instances"
                ],
                "summary": "Starts acc instance",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Instance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.AccWError"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.AccWError"
                        }
                    }
                }
            }
        },
        "/instance/{id}/stop": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Stops acc instance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instances"
                ],
                "summary": "Stops acc instance",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Instance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.AccWError"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.AccWError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.AccWError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "app.ExtraAccSettings": {
            "type": "object",
            "properties": {
                "adminPasswordIsEmpty": {
                    "type": "boolean"
                },
                "passwordIsEmpty": {
                    "type": "boolean"
                },
                "spectatorPasswordIsEmpty": {
                    "type": "boolean"
                }
            }
        },
        "app.InstancePayload": {
            "type": "object",
            "properties": {
                "acc": {
                    "$ref": "#/definitions/instance.AccConfigFiles"
                },
                "accExtraSettings": {
                    "$ref": "#/definitions/app.ExtraAccSettings"
                },
                "accWeb": {
                    "$ref": "#/definitions/instance.AccWebSettingsJson"
                },
                "id": {
                    "type": "string"
                },
                "is_running": {
                    "type": "boolean"
                },
                "path": {
                    "type": "string"
                },
                "pid": {
                    "type": "integer"
                }
            }
        },
...
        "app.SaveInstancePayload": {
            "type": "object",
            "properties": {
                "acc": {
                    "$ref": "#/definitions/instance.AccConfigFiles"
                },
                "accExtraSettings": {
                    "$ref": "#/definitions/app.ExtraAccSettings"
                },
                "accWeb": {
                    "$ref": "#/definitions/instance.AccWebSettingsJson"
                }
            }
        },
        "app.accWebInstanceLogs": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "logs": {
                    "type": "string"
                }
            }
        },
        "instance.AccConfigFiles": {
            "type": "object",
            "properties": {
                "assistRules": {
                    "$ref": "#/definitions/instance.AssistRulesJson"
                },
                "bop": {
                    "$ref": "#/definitions/instance.BopJson"
                },
                "configuration": {
                    "$ref": "#/definitions/instance.ConfigurationJson"
                },
                "entrylist": {
                    "$ref": "#/definitions/instance.EntrylistJson"
                },
                "event": {
                    "$ref": "#/definitions/instance.EventJson"
                },
                "eventRules": {
                    "$ref": "#/definitions/instance.EventRulesJson"
                },
                "settings": {
                    "$ref": "#/definitions/instance.SettingsJson"
                }
            }
        },
        "instance.AccWebAdvWindowsSettingsJson": {
            "type": "object",
            "properties": {
                "coreAffinity": {
                    "type": "integer"
                },
                "cpuPriority": {
                    "type": "integer"
                },
                "enableWindowsFirewall": {
                    "type": "boolean"
                }
            }
        },
        "instance.AccWebSettingsJson": {
            "type": "object",
            "properties": {
                "advWindowsCfg": {
                    "$ref": "#/definitions/instance.AccWebAdvWindowsSettingsJson"
                },
                "autoStart": {
                    "type": "boolean"
                },
                "enableAdvWindowsCfg": {
                    "type": "boolean"
                },
                "enableGlobalBanlist": {
                    "type": "boolean"
                },
                "enableGlobalEntrylist": {
                    "type": "boolean"
                }
            }
        },
        "instance.AssistRulesJson": {
            "type": "object",
            "properties": {
                "configVersion": {
                    "type": "integer"
                },
                "disableAutoClutch": {
                    "type": "integer"
                },
                "disableAutoEngineStart": {
                    "type": "integer"
                },
                "disableAutoGear": {
                    "type": "integer"
                },
                "disableAutoLights": {
                    "type": "integer"
                },
                "disableAutoPitLimiter": {
                    "type": "integer"
                },
                "disableAutoWiper": {
                    "type": "integer"
                },
                "disableAutosteer": {
                    "type": "integer"
                },
                "disableIdealLine": {
                    "type": "integer"
                },
                "stabilityControlLevelMax": {
                    "type": "integer"
                }
            }
        },
        "instance.BopJson": {
            "type": "object",
            "properties": {
                "configVersion": {
                    "type": "integer"
                },
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/instance.BopSettings"
                    }
                }
            }
        },
        "instance.BopSettings": {
            "type": "object",
            "properties": {
                "ballastKg": {
                    "type": "integer"
                },
                "carModel": {
                    "type": "integer"
                },
                "restrictor": {
                    "type": "integer"
                },
                "track": {
                    "type": "string"
                }
            }
        },
...
        "instance.ConfigurationJson": {
            "type": "object",
            "properties": {
                "configVersion": {
                    "type": "integer"
                },
                "lanDiscovery": {
                    "type": "integer"
                },
                "maxConnections": {
                    "type": "integer"
                },
                "publicIP": {
                    "type": "string"
                },
                "registerToLobby": {
                    "type": "integer"
                },
                "tcpPort": {
                    "type": "integer"
                },
                "udpPort": {
                    "type": "integer"
                }
            }
        },
...
        "instance.EntrylistJson": {
            "type": "object",
            "properties": {
                "configVersion": {
                    "type": "integer"
                },
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/instance.EntrySettings"
                    }
                },
                "forceEntryList": {
                    "type": "integer"
                }
            }
        },
        "instance.EventJson": {
            "type": "object",
            "properties": {
                "ambientTemp": {
                    "type": "integer"
                },
                "cloudLevel": {
                    "type": "number"
                },
                "configVersion": {
                    "type": "integer"
                },
                "isFixedConditionQualification": {
                    "type": "integer"
                },
                "metaData": {
                    "type": "string"
                },
                "postQualySeconds": {
                    "type": "integer"
                },
                "postRaceSeconds": {
                    "type": "integer"
                },
                "preRaceWaitingTimeSeconds": {
                    "type": "integer"
                },
                "rain": {
                    "type": "number"
                },
                "sessionOverTimeSeconds": {
                    "type": "integer"
                },
                "sessions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/instance.SessionSettings"
                    }
                },
                "simracerWeatherConditions": {
                    "type": "integer"
                },
                "track": {
                    "type": "string"
                },
                "trackTemp": {
                    "type": "integer"
                },
                "weatherRandomness": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}